+++ react/BazaarvoicePixel.tsx
@@ -7,7 +7,7 @@ export function handleEvents(e: PixelMessage) {
       const data = e.data
       const transactionData = {
         orderId: data.orderGroup,
-        total: data.transactionSubtotal,
+        total: data.transactionSubtotal - Math.abs(data.transactionDiscounts),
         currency: data.currency,
         tax: data.transactionTax,
         shipping: data.transactionShipping,
@@ -15,9 +15,12 @@ export function handleEvents(e: PixelMessage) {
         state: data.visitorAddressState,
         email: data.visitorContactInfo[0],
         nickname: data.visitorContactInfo[1],
+        discount: Math.abs(data.transactionDiscounts),
         items: data.transactionProducts.map(product => {
           return {
+            productId: product.id,
             sku: product.sku,
+            discount: product.originalPrice - product.price,
             quantity: product.quantity,
             name: product.name,
             price: product.price,